fix: test, remove debugger, fix cv filter
